refactor(billets-app): :recycle: refactored deep linking
