+++ apps/billets-app/src/lib/navigations/index.ts
@@ -1,1 +1,2 @@
 export * from './constants'
+export * from './deep-linking'
